feat(assembler): 重构字节码格式并添加新功能
- 更新字节码格式，增加函数本地变量和栈大小信息
- 添加类型信息到字节码中
- 优化字节码读取和写入逻辑
-重构虚拟机栈帧结构
- 更新函数解析和字节码生成逻辑

+++ assembler/src/main/java/io/github/leo1024/otrvm/conf/Type.java
@@ -13,12 +13,12 @@ public enum Type {
     STRING("str", 7, 8);
 
     private final String symbol;
-    private final byte id;
+    private final int id;
     private final int length;
 
     Type(String symbol, int code, int length) {
         this.length = length;
-        this.id = (byte) code;
+        this.id = code;
         this.symbol = symbol;
     }
 
@@ -37,7 +37,7 @@ public enum Type {
         return this != STRING;
     }
 
-    public byte getId() {
+    public int getId() {
         return id;
     }
